refactor(tee): centralize remote attestation verification
