+++ crates/blueprint-remote-providers/src/attestation/gcp.rs
@@ -6,20 +6,12 @@
 //! over an HTTPS endpoint the deployed workload exposes (the agent forwards the
 //! launcher token, optionally challenged with our nonce).
 //!
-//! Verification is a real JWT check against Google's published JWKS for the
-//! `confidentialcomputing.googleapis.com` issuer, plus claim policy (audience,
-//! nonce, image digest, debug-mode, freshness). See [`super::jwt`].
+//! Verification is delegated to
+//! [`blueprint_tee::attestation::providers::gcp_confidential::GcpConfidentialVerifier`].
 
-use super::jwt::{JwksSource, verify_jwt_attestation};
-use super::{AttestationError, AttestationPolicy, TeeAttestationGate, VerifiedTeeDeployment};
+use super::{AttestationError, AttestationPolicy, TeeAttestationGate, VerifiedAttestation};
 use crate::core::remote::CloudProvider;
-
-/// Issuer of GCP Confidential Space attestation tokens.
-pub const GCP_CONFIDENTIAL_ISSUER: &str = "https://confidentialcomputing.googleapis.com";
-
-/// Google's JWKS endpoint for Confidential Space attestation tokens.
-pub const GCP_CONFIDENTIAL_JWKS_URL: &str =
-    "https://confidentialcomputing.googleapis.com/.well-known/jwks";
+use blueprint_tee::attestation::providers::gcp_confidential::GcpConfidentialVerifier;
 
 /// Default in-VM path (over HTTPS the workload re-exposes) for the attestation
 /// token. The port/path is a convention the deployed agent honours; callers can
@@ -29,11 +21,9 @@ const DEFAULT_TOKEN_PATH: &str = "/.well-known/attestation-token";
 /// Gate for GCP Confidential Space.
 pub struct GcpConfidentialSpaceGate {
     http: reqwest::Client,
-    jwks: JwksSource,
+    verifier: GcpConfidentialVerifier,
     /// Path on the VM endpoint that returns the OIDC attestation token.
     token_path: String,
-    /// Allowed token issuers.
-    allowed_issuers: Vec<String>,
 }
 
 impl GcpConfidentialSpaceGate {
@@ -41,15 +31,14 @@ impl GcpConfidentialSpaceGate {
     pub fn new() -> Self {
         Self {
             http: super::default_http_client(),
-            jwks: JwksSource::new(GCP_CONFIDENTIAL_JWKS_URL),
+            verifier: GcpConfidentialVerifier::new(),
             token_path: DEFAULT_TOKEN_PATH.to_string(),
-            allowed_issuers: vec![GCP_CONFIDENTIAL_ISSUER.to_string()],
         }
     }
 
     /// Override the JWKS URL (used by tests to point at a local mock).
     pub fn with_jwks_url(mut self, url: impl Into<String>) -> Self {
-        self.jwks = JwksSource::new(url);
+        self.verifier = self.verifier.with_jwks_url(url);
         self
     }
 
@@ -61,7 +50,7 @@ impl GcpConfidentialSpaceGate {
 
     /// Override allowed issuers (used by tests).
     pub fn with_allowed_issuers(mut self, issuers: Vec<String>) -> Self {
-        self.allowed_issuers = issuers;
+        self.verifier = self.verifier.with_allowed_issuers(issuers);
         self
     }
 }
@@ -89,18 +78,18 @@ impl TeeAttestationGate for GcpConfidentialSpaceGate {
             .send()
             .await
             .map_err(|e| AttestationError::Fetch {
-                provider: CloudProvider::GCP,
+                provider: "GCP".to_string(),
                 reason: format!("token fetch from {url} failed: {e}"),
             })?;
         if !resp.status().is_success() {
             return Err(AttestationError::Fetch {
-                provider: CloudProvider::GCP,
+                provider: "GCP".to_string(),
                 reason: format!("token endpoint returned HTTP {}", resp.status()),
             });
         }
         let body = super::read_body_capped(resp, super::MAX_TOKEN_BYTES, |reason| {
             AttestationError::Fetch {
-                provider: CloudProvider::GCP,
+                provider: "GCP".to_string(),
                 reason,
             }
         })
@@ -112,21 +101,12 @@ impl TeeAttestationGate for GcpConfidentialSpaceGate {
         &self,
         evidence: &[u8],
         policy: &AttestationPolicy,
-    ) -> Result<VerifiedTeeDeployment, AttestationError> {
+    ) -> Result<VerifiedAttestation, AttestationError> {
         let token = std::str::from_utf8(evidence).map_err(|e| AttestationError::Malformed {
-            provider: CloudProvider::GCP,
+            provider: "GCP".to_string(),
             reason: format!("token is not UTF-8: {e}"),
         })?;
-        let issuers: Vec<&str> = self.allowed_issuers.iter().map(String::as_str).collect();
-        verify_jwt_attestation(
-            CloudProvider::GCP,
-            token,
-            &self.jwks,
-            &issuers,
-            policy,
-            &self.http,
-        )
-        .await
+        self.verifier.verify_token(token, policy).await
     }
 }
