+++ crates/tee/Cargo.toml
@@ -52,6 +52,11 @@ gcp_auth = { workspace = true, optional = true }
 reqwest = { workspace = true, features = ["json"], optional = true }
 base64 = { workspace = true, features = ["std"], optional = true }
 uuid = { workspace = true, features = ["v4"], optional = true }
+jsonwebtoken = { workspace = true, optional = true }
+coset = { workspace = true, optional = true }
+ciborium = { workspace = true, optional = true }
+p384 = { workspace = true, optional = true }
+x509-parser = { workspace = true, optional = true }
 
 [dev-dependencies]
 tokio = { workspace = true, features = ["full"] }
@@ -69,15 +74,24 @@ std = ["blueprint-core/std", "blueprint-std/std"]
 
 ## Enable AWS Nitro Enclave runtime backend and attestation verification.
 ## Adds `aws-sdk-ec2` and `aws-config` for EC2 instance management.
-aws-nitro = ["dep:aws-sdk-ec2", "dep:aws-config", "dep:base64", "dep:uuid"]
+aws-nitro = [
+    "dep:aws-sdk-ec2",
+    "dep:aws-config",
+    "dep:base64",
+    "dep:uuid",
+    "dep:coset",
+    "dep:ciborium",
+    "dep:p384",
+    "dep:x509-parser",
+]
 
 ## Enable Azure SEV-SNP / SKR runtime backend and attestation verification.
 ## Adds `reqwest` for ARM REST API and MAA token retrieval.
-azure-snp = ["dep:reqwest", "dep:base64", "dep:uuid"]
+azure-snp = ["dep:reqwest", "dep:base64", "dep:uuid", "dep:jsonwebtoken"]
 
 ## Enable GCP Confidential Space runtime backend and attestation verification.
 ## Adds `gcp-auth` for service account auth and `reqwest` for Compute Engine API.
-gcp-confidential = ["dep:gcp_auth", "dep:reqwest", "dep:uuid"]
+gcp-confidential = ["dep:gcp_auth", "dep:reqwest", "dep:uuid", "dep:jsonwebtoken"]
 
 ## Enable Intel TDX attestation verification
 tdx = []
@@ -90,6 +104,9 @@ all-providers = ["aws-nitro", "azure-snp", "gcp-confidential", "tdx", "sev-snp"]
 
 ## Enable test utilities (e.g., VerifiedAttestation::new_for_test)
 test-utils = []
+
+## Allow policy-controlled debug TEE acceptance in tests only.
+testing = []
 
 [package.metadata.docs.rs]
 all-features = true
