+++ crates/tee/src/attestation/providers/gcp_confidential.rs
@@ -1,17 +1,23 @@
 //! GCP Confidential Space attestation verifier.
 //!
-//! Currently validates GCP Confidential Space attestation structurally:
-//! - Measurement comparison
-//! - Debug mode detection
-//!
-//! **Limitation:** Token signature verification, workload identity validation,
-//! and machine family TEE type derivation are not yet implemented. These require
-//! provider-specific dependencies that will be added in a future release.
+//! `verify_token` performs cryptographic verification of the Confidential Space
+//! OIDC attestation token against Google's JWKS plus nonce/audience/workload
+//! policy checks. The synchronous [`AttestationVerifier`] implementation remains
+//! a structural check for already-materialized reports.
 
-use crate::attestation::report::AttestationReport;
+use crate::attestation::providers::jwt::{JwksSource, verify_jwt_attestation};
+use crate::attestation::report::{AttestationFormat, AttestationReport};
 use crate::attestation::verifier::{AttestationVerifier, VerifiedAttestation};
+use crate::attestation::{AttestationError, AttestationPolicy};
 use crate::config::TeeProvider;
 use crate::errors::TeeError;
+
+/// Issuer of GCP Confidential Space attestation tokens.
+pub const GCP_CONFIDENTIAL_ISSUER: &str = "https://confidentialcomputing.googleapis.com";
+
+/// Google's JWKS endpoint for Confidential Space attestation tokens.
+pub const GCP_CONFIDENTIAL_JWKS_URL: &str =
+    "https://confidentialcomputing.googleapis.com/.well-known/jwks";
 
 /// Verifier for GCP Confidential Space attestation tokens.
 pub struct GcpConfidentialVerifier {
@@ -19,6 +25,9 @@ pub struct GcpConfidentialVerifier {
     pub expected_measurement: Option<String>,
     /// Whether to allow debug-mode VMs.
     pub allow_debug: bool,
+    http: reqwest::Client,
+    jwks: JwksSource,
+    allowed_issuers: Vec<String>,
 }
 
 impl GcpConfidentialVerifier {
@@ -27,6 +36,9 @@ impl GcpConfidentialVerifier {
         Self {
             expected_measurement: None,
             allow_debug: false,
+            http: default_http_client(),
+            jwks: JwksSource::new(GCP_CONFIDENTIAL_JWKS_URL),
+            allowed_issuers: vec![GCP_CONFIDENTIAL_ISSUER.to_string()],
         }
     }
 
@@ -40,6 +52,44 @@ impl GcpConfidentialVerifier {
     pub fn allow_debug(mut self, allow: bool) -> Self {
         self.allow_debug = allow;
         self
+    }
+
+    /// Override the JWKS URL, primarily for tests.
+    pub fn with_jwks_url(mut self, url: impl Into<String>) -> Self {
+        self.jwks = JwksSource::new(url);
+        self
+    }
+
+    /// Override allowed token issuers, primarily for tests.
+    pub fn with_allowed_issuers(mut self, issuers: Vec<String>) -> Self {
+        self.allowed_issuers = issuers;
+        self
+    }
+
+    /// Verify a Confidential Space OIDC attestation token cryptographically.
+    pub async fn verify_token(
+        &self,
+        token: &str,
+        policy: &AttestationPolicy,
+    ) -> Result<VerifiedAttestation, AttestationError> {
+        let mut policy = policy.clone();
+        if self.allow_debug {
+            policy.allow_debug = true;
+        }
+        if policy.expected_measurement.is_none() {
+            policy.expected_measurement = self.expected_measurement.clone();
+        }
+        let issuers: Vec<&str> = self.allowed_issuers.iter().map(String::as_str).collect();
+        verify_jwt_attestation(
+            TeeProvider::GcpConfidential,
+            AttestationFormat::GcpConfidentialToken,
+            token,
+            &self.jwks,
+            &issuers,
+            &policy,
+            &self.http,
+        )
+        .await
     }
 }
 
@@ -81,7 +131,7 @@ impl AttestationVerifier for GcpConfidentialVerifier {
         }
 
         tracing::debug!(
-            "structural validation only — cryptographic signature verification requires token validation"
+            "structural validation only; call verify_token() for cryptographic JWT verification"
         );
 
         Ok(VerifiedAttestation::new(
@@ -93,4 +143,11 @@ impl AttestationVerifier for GcpConfidentialVerifier {
     fn supported_provider(&self) -> TeeProvider {
         TeeProvider::GcpConfidential
     }
+}
+
+fn default_http_client() -> reqwest::Client {
+    reqwest::Client::builder()
+        .timeout(std::time::Duration::from_secs(15))
+        .build()
+        .unwrap_or_default()
 }
